+++ Cafe_App/Data/Masa.cs
@@ -13,9 +13,9 @@ public partial class Masa
 
     public int Kapasite { get; set; }
 
-    public decimal Tutar { get; set; }
+    public float Tutar { get; set; }
 
-    public decimal OdenenTutar { get; set; }
+    public float OdenenTutar { get; set; }
 
     public int PersonelId { get; set; }
